+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Fix
 - Main character skin to local player not correctly applied
+- NPC name tag shown in a story